Make joining the server async.
All logic of onPlayerJoin was handled synchronously, which lead to
weird behaviour:
 - the whole server probably freezed when a player entered
 - we could not send messages while setting up the world
 - the initial world was not loaded at all and the player was
   in a completely empty world without any blocks
 - the server complained about not being able to keep up

There is now a new method execSync, which runs lambda expressions
synchronously in the main thread and waits for it to finish.  With
this method it is possible to make computations async with few
exceptions.

Note: This commit is a WIP.  It works but especially the pregeneration
of worlds could have a race condition (and I don't know whether it is
needed at all).

We now make sure to only send updates for players for which a game has
already been registered.  Otherwise, the async nature would try to
send updates for players whose game is still being set up.