Set blocks by region, delay world generation, clean data on setup
The blocks we want to have in the world were added in the same
tick (probably) as the world generation.  Delay it by two ticks to get
rid of the ConcurrentModificationException
(see https://github.com/minecraft-saar/documents/issues/44)
Don't know why this works and I could not finde good documentation on
this topic.

When setup.sh is run, delete server and replay_server to always have a
clean state.

We use setRegion instead of setBlock to reduce code size.  No
functional change.